Break out iac test command
